Another type issue in method signature

+++ assign04/symbol.cpp
@@ -24,7 +24,7 @@ long Symbol::get_offset() {
     return m_offset;
 }
 
-struct Symbol *symbol_create(const char* name, Type* type, int kind, int offset) {
+struct Symbol *symbol_create(const char* name, Type* type, int kind, long offset) {
     Symbol *symbol = new Symbol();
     symbol->m_name = name;
     symbol->m_type = type;
